server/requirements.txt
asgiref==3.3.4
attrs==21.2.0
awsebcli==3.20.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blessed==1.18.1
botocore==1.20.97
bs4==0.0.1
cached-property==1.5.2
cement==2.8.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.3
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
Django==3.2.4
django-braces==1.14.0
django-cors-headers==3.7.0
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
ecdsa==0.17.0
Flask==2.0.1
Flask-Cors==3.0.10
future==0.16.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jplephem==2.15
jsonfield==3.1.0
jsonschema==3.2.0
jwcrypto==0.9.1
langdetect==1.0.9
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.20.3
oauth2==1.9.0.post1
oauthlib==3.1.1
paramiko==2.7.2
pathspec==0.5.9
pyasn1==0.4.8
pycparser==2.20
pymongo==3.11.4
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
semantic-version==2.8.5
sgp4==2.19
six==1.14.0
skyfield==1.39
social-auth-app-django==4.0.0
social-auth-core
soupsieve==2.2.1
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
typing-extensions==3.10.0.0
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.59.0
Werkzeug==2.0.1
whois==0.9.13
wrapt==1.12.1
zipp==3.4.1
python-dotenv==0.18.0